dock.go improvements
The (camlistored builder + UI resources bundler) was a piece of shell
run in a docker container. That piece of shell is replaced with the
misc/docker/server/build-camlistore-server.go go program. This
program is still run within a container of the camlistore/go image.
It is added "on the fly" as a resource to the camlistore/go image
as a mounted resource.

In addition, build-camlistore-server.go can take a local directory as
the camlistore source (instead of fetching a tarball at the given
revision). dock.go has been adapted to take advantage of that.

Change-Id: If32b523d850f6c932016f18231bb1b1a283e50d3