devcam: move modtime check after cwd check
checkCamliSrcRoot was meant to give a nice error msg if devcam is
invoked from outside the Camlistore source tree. However, it was never
called because the modtime check (checkModtime), which itself needs to
be run from that same location, was called first. So we were always
getting the terse "open dev/devcam: no such file or directory" error
message instead.

Fixes #589

Change-Id: I13a1bb3819217ffbaf727fa6154f077dd8babd49