misc/docker: build tarball of binaries for releases
This change:

1) adds misc/docker/go/darwin/Dockerfile to create a docker image with
Go setup for cross-compiling to darwin
2) adds misc/docker/release/build-binaries.go, which works similarly to
misc/docker/server/build-camlistore-server.go. The difference is it
builds all binaries with make.go instead of just camlistored, and
camlistored is hence built with all resources embedded.
3) adds the --build_binaries flag to dock.go to make use of the above.

Maybe 3) should be done in a new file (out of dock.go), but it is
convenient for now for code reuse.

Issue #538

Change-Id: I8ecd50d118a9ff5c3f53812cac4d7a81a0053b4f