pkg/video/thumbnail: for making a still thumbnail from a video
This package is built around 3 components: a Service, a Thumbnailer, and
an http handler.
When the Service generates a thumbnail, it creates a Thumbnailer (based
on ffmpeg by default), and it starts an ephemeral http server. The
Thumbnailer runs a thumbnailing process that reads the video file served
by the ephemeral server, and outputs a thumbnail image.

Not plugged with Camlistore yet.

Commiter note: submitting as is from files taken from
https://camlistore-review.googlesource.com/3582 to conserve authorship.
I'll tweak the package if needed in subsequent commits.

Change-Id: I285a7694d1814d52e1248caf4eb501810eed89c1