ui/server_connection: fix opt_fail handling + long overdue cleanups
opt_fail as an optional argument didn't work in the cases where the
func would also receive an event as argument. Because when opt_fail is
not provided, the arguments get shifted and the event value would end up
in the opt_fail variable, or vice-versa.
So in these cases, I switched to the more robust
func({success: success, fail: fail}, event) syntax

As a result, safeFail wasn't that useful anymore, and it was incorrectly
bound in some cases, so I removed it.

I also removed a few intermediate function calls which made reading
unnecessarily more complicated.

Change-Id: Ie70694a26e52f302837bbf6ffc16b0a0bb77ad7e