importer, twitter: fix deleted account type, differentiate account and root titles
When being deleted the importer account node was changing from an
"importerAccount" to an "importer-deleted" nodeType. I fixed it to
"importerAccount-deleted" for consistency.

Also, both the importer account node and the import root node ended up
with the "username's Twitter Account" title which was looking confusing
in the UI. I fixed the root node title to be "username's Twitter Data"
instead.

Change-Id: I335856ca92aca897a05d94a5f3e7d438b00ecc3e