pkg/rollsum: fix uint8 overflow
As drop is a uint8, and charOffset is a const, drop + charOffset is a
uint8, which overflows if the result is > 255.

The original code
(at https://github.com/bup/bup/blob/master/lib/bup/bupsplit.c):
  r->s2 += r->s1 - (BUP_WINDOWSIZE * (drop + ROLLSUM_CHAR_OFFSET));
does not have this problem because drop in this case is implicitly
converted to unsigned before the arithmetic is applied, according the
the C 'usual arithmetic conversions'.

This change thus converts drop to a uint32 before the addition.

pkg/schema/filewriter_test.go: tests repaired

By changing the chunking boundaries, this modification to rollsum
breaks the tests for filewriter. The number and boundaries of chunks
changes and the fileMap describing the file changes accordingly.

Fixes #611

Change-Id: Ica3140a194b1218921e6ae0ac3206ccd0b93fa72