website: convert docs pages to markdown
This includes all of the top-level files in the docs folder except for
status and terms, which are going to be a little trickier to convert.

This includes no editorial changes.  There are a few minor stylistic
changes (like the auth value table in server-config converted into a
list).  Text has also been mostly wrapped at 80 columns, which accounts
for a lot of the changes as well.

Change-Id: I3782158f05d2b3c72e6cb44bff92c46ccc231a96