pkg/client: be stricter when rewriting URLs
url.Parse in Go 1.5 now fails with the scheme://IPv6:/foo form (trailing
colon). But it still works when the hostname is an actual hostname, or
an IPv4, which seems inconsistent.
This CL therefore not only accepts this new failure (which comes with Go
1.5) but also considers any host with a trailing colon as an invalid URL
and refuses to rewrite it.

There probably are other places in Camlistore where we should the same.

See https://github.com/golang/go/issues/12200

Change-Id: Ib9ea95a71013b5b53f2fd99415015ec916bb4d9d