UI: add to set when dropped is a dup but not yet a member of the set
Follow-up to 82c6feb61a101b8f6bdaf17523ecf1d30d94cadd and builds upon the work of Mathieu Lonjaret's Change 5345.

When viewing a set, and the dropped file is a dup for an existing permanode('s content), we still need to add the permanode to the set if it is not already a member of it. The above commit overlooked that point.

Fixes #622

Change-Id: I694e583625886a7bf164ab9fd5bf8874a7cc2b23