pkg/serverinit: use sortedStorageAt for diskpacked
This makes diskpacked index type/location be set the same way it is done for
blobpacked, that is, its type will be the same as the camlistore index's
is. e.g. if the camli index is in MySQL, then the diskpacked index will
be another database in MySQL too.

Additionally, for added clarity wrt to the changes in
https://camlistore-review.googlesource.com/5496 , if we use a file based
indexer, the diskpacked index filename will now be
blobPath/diskpacked-index.kvtype (instead of blobPath/index.kvtype).

Change-Id: I7a9eb5c9caf07f62732ba2771308d351ec37b814