pkg/images/resize: disable TestCompareResizeToHalveInplace for now
It's more important for now to be able to monitor for other failures in
the Camlistore repo than leaving this test on.

After having looked into the issue, I know a little more but not enough
for a definitive conclusion or fix.
It's all added as comments so we don't start from scratch again later.

Also, to make up for the loss of this test, I'm adding two other tests,
TestCompareOriginalToHalveInPlace, and TestCompareOriginalToResized,
that are similar enough to catch other problems in the same context I
believe.

Context: issue #635

Change-Id: I80a30f544fb9ab2e7c69118ef8509d3d687c2d27