pkg/client: add UploadFile
Most of the code related to uploading a file with a client is actually
locked into the camput code. And uploading a file is something the
scanning cabinet app needs to do.

Refactoring the camput file code into pkg/client, and cleaning up
pkc/client is a lot of work. Therefore I propose with this change to
first introduce the minimum I need in pkg/client that I can use in
the scanning cabinet app.

Change-Id: Ie8690e66f0332be76a3a41fa060db410d01798e2