pkg/search: don't run a speculative search more than once a time
When new blobs arrive we re-run open search queries (those requested
by an active websocket client) in case they changed, and then push the
new result to the browser if they did change.

But when uploading lots of data if the browser has a big search query
active, it was possible to start running the same query in many
goroutines over the same data. Don't do that.

Verified that this was the case with some temporary logging in the
already-refreshing case.

Updates camlistore/camlistore#519