pkg/serverinit: fix blobpacked_index genconfig for GCS
The b.sortedStorageAt call had an empty filePrefix argument, which
means it would fail if the current index was not a DBMS (but a
file-based DB).

Also, this change adds a default "blobpacked_index" database name value
for the case when a user wants a DBMS index, but hasn't configured a
DBNames map.

Issue #693

Change-Id: Ie4386e789542c006530983e20bccc7a564bc4ee3