pkg/httputil: fix checkSystemRoots for Go1.6
Something changed in x509 in Go1.6 so that when verifying an empty Cert,
one now gets the errNotParsed (x509: missing ASN.1 contents; use
ParseCertificate) error, instead of the SystemRootsError.

Since we rely on getting the SystemRootsError to find out that we're on
CoreOS, we now need to verify a non empty cert, when performing this
check.

Fixes issue #705

Change-Id: I40a4f9aa4ef49bbdd3b7b15b127f890e26f31de4