importer/picasa: rm anyErr because never set
anyErr is supposed to track if at any point there was an error in the
importing, so we know whether to mark it as fully successful.
As opposed to e.g. with foursquare, all errors encountered by the picasa
importer are propagated up the call chain, so:

	if err := r.importAlbums(); err != nil {
		return err
	}

is enough to ensure that we set AcctAttrCompletedVersion only if the run
was successful.

Change-Id: I8c5dbbf853b84391af70372f81c08588adb76884