website: do a full cloning of camlistore.org on startup
We used to have --depth=1 when cloning camlistore.org on startup, for
efficiency reasons.
But not having all the commits locally caused problems when pushing to
github, because it would see the remote had commits that we didn't.
So we're now doing a full cloning.

It looks like it's now taking ~4s (as opposed to ~2.5s before), and
consuming ~50MB diskspace (as opposed to 35MB before).

Plus a few bugfixes.

Change-Id: If7dbae1d3119d8b3336fb7d735f6bd0ba7606fc6