convert protocol docs to markdown
Update links to reference the website URLs rather than gerrit.  There is
some more cleanup that could be done with the docs (for example, using
tables for the request and response keys), but this is focused on just
basic rendering.

Git is treating these as file deletions and additions... apparently
there were too many changes to treat them as renames, though that's what
they really are.

Refs #720

Change-Id: Ibca17b78e1412c049843236bfbc8f2a9e368bb23