website: use custom markdown options
This is all so that we can enable auto creating header IDs based on the
text of the header.  For example "# Foo" is converted to:

  <h1 id="foo">Foo</h1>

I'm not really sure why this isn't a default option, but whatever.  It's
also unfortunate that we have to copy all of the flags and extensions
from MarkdownCommon, but at least this will make it easier to add more
options in the future, I gues.

Change-Id: I65800c75560ad3b9a3fd0389b722c07ba11ee098