website: restore search-ui.txt with link to new page
The help page in the camlistore binary has a direct link for "Search bar
predicates" that links to this search-ui.txt on gerrit.  Restore this
file so that these links will still resolve, but with a pointer to the
new docs on camlistore.org.

Also update the help link for future builds of camlistore to point to
the correct URL.

Change-Id: Ib81ff820869d86e6e6f664e8b52c5f20d4eed4d3