serverinit: remove trailing commas causing invalid JSON
configHandler strips the key "_knownkeys" from all JSON objects in its
output using a regex on the marshaled JSON.  If this was the last key of
an object which contains multiples keys, this leaves the JSON in an
invalid state.  For example,

    {
        "foo": "bar",
        "_knownkeys": {}
    }

becomes:

    {
        "foo": "bar",
    }

The trailing comma here is illegal.  To fix this, we run one more regex
across the output removing any comma at the end of a line just before a
closing brace.  This is never valid in JSON, so is safe to remove
globally here.

Change-Id: I05f54e9f094fbe9cbc720f073b2b78ae80525106