misc/docker: replace CopyObject calls
storage.Client.CopyObject does not exist anymore. Replaced with CopyTo
calls.

Also removed checks based on the presence of /.dockerinit to know
whether we're on docker because apparently they're gone or something.
I'll investigate why and a replacement later; we want a new GCE docker
image asap and I'm pressed for time.

Change-Id: I6b3ebcb048f03bdf53db94b41fe96ddcb6ed7537