website: fix remaining old links to /docs/
interestingly, I also discovered that GitHub will properly resolve links
that are relative to the root path (e.g. "/doc/foo")...  it properly
rewrites these links to reference the named file relative to the
repository root.

also add a /doc/contributing.md file.  This will never be served from
the website, since we have a redirect handler in place, but is added so
that we don't have broken links when browsing the docs on GitHub.

Change-Id: I47517a5f6303065df02ac11771dcee56cd06a4b7