make.go: mirror assembly code too
pkg/fs/fs_test.go depends on bazil.org/fuse/syscallx, which in turn
depends on golang.org/x/sys/unix
We already had the assembly code vendored in for golang.org/x/sys/unix,
but we weren't mirroring it in make.go

This didn't show with 'go run make.go' because it only affects a test.
And it didn't show with 'go test ./pkg/fs' either, because the assembly
would then be found in GOPATH.
Only devcam test would detect it.

Change-Id: Ibb90a5fb986a3d0308ff9f1f8df8d5866cfec8c5