pkg/client: fail properly when no SearchRoot
When the discovery SearchRoot is empty (because e.g. the server has no
indexer), the root.Parse(disco.SearchRoot) call during doDiscovery will
not fail. This is fine, and is actually what we want, because we don't
want discovery to fail just because there's no SearchRoot.

However, we do want func (c *Client) SearchRoot() to return
ErrNoSearchRoot in that case, which is why c.searchRoot should be set to
"" empty as well.

Change-Id: I88cbfa4a1476ba1533ccd61511a72c35de6d7516