pkg/search: lock index for corpus creation
As soon as the (bs->index) sync handler is initialized, it starts
writing pending blobs to the index. However, so far the corpus
initialization has been left, unguarded, up to the search handler. Which
means lots of potential data races on the corpus fields between the sync
handler writing a blob to the index, and the search handler creating all
the pieces of the corpus itself.

This change fixes the issue by locking the index while the search
handler is creating the corpus.

This change does not entirely fix issue #750, as there seems to be yet
another data race going on, but it is related as the fixed race was
found thanks to the race trace at:
https://github.com/camlistore/camlistore/issues/750#issuecomment-241881987

Change-Id: I6de0165064a4f8934466f6a08710b4962f7b0256