website: clean up previous docker git container
All the camweb features depending on docker had been stuck for a couple
of weeks because one of the docker containers was wedged somehow. I
don't know exactly what was wrong, but I figure it can't hurt to add the
same kind of cleanup that we have for the demo blobserver.

So in this CL, we name each container running a git command, so we can
stop and remove said container the next time we run the same command.

Change-Id: I66592fbbde73ea30e4cee7477ada450e0c6a645e