camlistored/ui: implement swipe for left/right nav on mobile
Naive implementation for now. I imitated how things were setup for the
left and right keys, but maybe the listener should be on the image
container instead of on the index page? It seems to work anyway.

No animation when starting to swipe though, which means users have no
clue swipe exists/works until they actually try to do it for real.

Change-Id: If0baed3c8d3ff1f5d8886fca891a8ac40980f66e